Implement a new class type CreatureBehaviour to customize creature behaviours

+++ source/utils/ConfigManager.h
@@ -114,7 +114,7 @@ public:
     inline const std::map<const std::string, std::vector<const CreatureMood*> >& getCreatureMoodModifiers() const
     { return mCreatureMoodModifiers; }
     CreatureMoodLevel getCreatureMoodLevel(int32_t moodModifiersPoints) const;
-    inline int64_t getNbTurnsFuriousMax() const
+    inline int32_t getNbTurnsFuriousMax() const
     { return mNbTurnsFuriousMax; }
 
     inline double getMaxManaPerSeat() const
@@ -265,7 +265,7 @@ private:
     int32_t mCreatureMoodFurious;
     double mSlapDamagePercent;
     int64_t mTimePayDay;
-    int64_t mNbTurnsFuriousMax;
+    int32_t mNbTurnsFuriousMax;
     double mMaxManaPerSeat;
     double mClaimingWallPenalty;
     double mDigCoefGold;
